Implement UDP service
A thread is created which binds to the port and then waits for data.
The same socket is used for all datagrams so only one thread is
required to handle all UDP packets.

If a datagram is larger than the buffer then the data is truncated.
This means the limit is currently set to 1KB.